requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0


pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
xlrd>=2.0.1

google-genai>=0.2.0

python-dotenv>=1.0.0

requests>=2.31.0

gunicorn>=21.2.0
whitenoise>=6.6.0